refactor: replace handlePostSignIn with server action; write new userData and invite status via server actions in server components; remove inviteId atom
