+++ src/app/login/page.tsx
@@ -17,7 +17,12 @@ export default async function LoginPage({
       <div className="min-h-screen flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
         <div className="mt-10">
           <Suspense>
-            <AuthClient defaultScreen={''} redirect={searchParams?.redirect} />;
+            <AuthClient
+              defaultScreen={''}
+              redirect={searchParams?.redirect}
+              inviteId={searchParams?.invite_id}
+            />
+            ;
           </Suspense>
         </div>
       </div>
