+++ src/components/user-portal/ProfileMain.tsx
@@ -1,12 +1,7 @@
 'use client';
 
-import { useState, useEffect } from 'react';
-
 import { motion } from 'framer-motion';
 import RenderCardThumbnail from '@/components/user-portal/CardThumbnail';
-import { useAtom } from 'jotai';
-import AlertModal from '@/components/common/AlertModal';
-import { AlertModalType } from '@/types/AlertModalType';
 import {
   ICards,
   UserData,
@@ -14,10 +9,6 @@ import {
   emptyOwnedCard,
 } from '@/types/User.type';
 import ProfileHeader from '@/components/user-portal/ProfileHeader';
-import { useAuthenticator } from '@aws-amplify/ui-react';
-import { inviteIdAtom } from '@/states/userStore';
-import { fetchUserAttributes } from 'aws-amplify/auth';
-import handlePostSignIn from '@/app/utils/auth/handlePostSignIn';
 
 //motion variants to animate the team bars
 const variants = {
@@ -34,99 +25,9 @@ export default function ProfileMain({ userData }: { userData: UserData }) {
   const guestCardsData =
     userData.guest_cards === null ? [emptyGuestCard] : userData.guest_cards;
 
-  const [inviteId] = useAtom(inviteIdAtom);
-  const [inviteStatus, setInviteStatus] = useState<AlertModalType | null>();
-  const [hasSetState, setHasSetState] = useState(false);
-
   const acceptedGuestCards = guestCardsData.filter(
     (card) => card.status === 'accepted',
   );
-
-  const { user, route } = useAuthenticator((context) => [
-    context.user,
-    context.route,
-  ]);
-
-  // This useEffect hook runs whenever the inviteData changes
-  // Listen for the sign-in event after user verifies their email and signs in
-  useEffect(() => {
-    if (user && route === 'authenticated') {
-      fetchUserAttributes()
-        .then((userAttributes) => {
-          handlePostSignIn(userAttributes, inviteId)
-            .then((inviteData) => {
-              // Get the value of 'hasShownModal' from local storage
-              const hasShownModal = localStorage.getItem('hasShownModal');
-              console.log('hasShownModal %s', hasShownModal);
-              // If the modal has not been shown and there is valid invite data
-              if (
-                !hasShownModal &&
-                inviteData &&
-                inviteData.invitation.invite_id &&
-                hasSetState === false
-                // inviteData.invitation.invite_status !== 'SUCCESS'
-              ) {
-                let inviteTitle: null | string = null;
-                let inviteMessage: null | string = null;
-
-                // Set the invite message based on the invite status
-                if (inviteData.invitation.invite_status === 'REVOKED') {
-                  inviteTitle = 'Card Access Revoked';
-                  inviteMessage =
-                    'We understand you were previously granted access to this card. However, the card owner has chosen to revoke your access privileges. We apologize for any inconvenience this may cause.';
-                } else if (inviteData.invitation.invite_status === 'EXPIRED') {
-                  inviteTitle = 'Invitation Expired';
-                  inviteMessage =
-                    'The access you were granted to this card has now expired. The card owner may choose to reinstate your access if they wish. Thank you for your understanding.';
-                } else if (
-                  inviteData.invitation.invite_status === 'UNEXPECTED_STATUS' ||
-                  inviteData.invitation.invite_status === 'NOT_FOUND'
-                ) {
-                  inviteTitle = 'Invitation Error';
-                  inviteMessage =
-                    "Oops! We encountered an issue processing your invitation. It's possible the invitation doesn't exist or there was a problem on our end. Please double-check the invitation details or contact the card owner for assistance.";
-                } else if (
-                  inviteData.invitation.invite_status === 'ALREADY_ACCEPTED'
-                ) {
-                  inviteTitle = 'Invitation Already Accepted';
-                  inviteMessage =
-                    'This invitation has already been redeemed. If you believe this is an error, please contact our support team for assistance.';
-                } else if (inviteData.invitation.invite_status === 'SUCCESS') {
-                  inviteTitle = 'Invitation Success';
-                  inviteMessage =
-                    'You have successfully accepted the invitation to access this card. You can now view and manage this card from your dashboard.';
-                } else if (
-                  inviteData.invitation.invite_status === 'SUCCESS_OWNER_SET'
-                ) {
-                  inviteTitle = 'Registration Success';
-                  inviteMessage =
-                    'Thank you for registering! You have successfully claimed your card. You can now view and manage it from your dashboard. If you have any questions, feel free to contact our support team.';
-                }
-
-                // Set the invite status state with the title and message
-                setInviteStatus({
-                  title: inviteTitle,
-                  textBody: inviteMessage,
-                });
-                // Store a flag in local storage to indicate that the modal has been shown
-                localStorage.setItem('hasShownModal', 'true');
-                setHasSetState(true);
-              }
-            })
-            .catch((err) => {
-              console.error('Error in post sign-in:', err);
-            });
-        })
-        .catch((err) => {
-          console.error('Error fetching user attributes:', err);
-        });
-    }
-  }, [route, inviteId, user]);
-
-  // Function to close the modal by setting the invite status to null
-  const closeModal = () => {
-    setInviteStatus(null);
-  };
 
   // ADD SKELETON
   return (
@@ -146,13 +47,6 @@ export default function ProfileMain({ userData }: { userData: UserData }) {
             animate="animate"
             className="flex flex-col items-center mt-4 md:pt-7"
           >
-            {inviteStatus && (
-              <AlertModal
-                title={inviteStatus.title}
-                textBody={inviteStatus.textBody}
-                onClose={closeModal}
-              />
-            )}
             <motion.div
               variants={variants}
               className="overflow-hidden w-full max-w-[1030px] mb-4 text-primary bg-cardsDark shadow-lg rounded-10 flex flex-col px-2 md:px-4 py-8"
